+++ main.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
 
-	testutils "github.com/HewlettPackard/hpegl-vmaas-terraform-resources/pkg/test-utils"
+	testutils "github.com/HewlettPackard/hpegl-vmaas-terraform-resources/internal/test-utils"
 	libUtils "github.com/hewlettpackard/hpegl-provider-lib/pkg/utils"
 )
